+++ CHANGELOG.md
@@ -1,5 +1,51 @@
 # 本版本更新包含以下内容：
 
+## 主机安全(cwp) 版本：2018-02-28
+
+### 第 17 次发布
+
+发布时间：2021-06-14 08:10:35
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增接口：
+
+* [DescribeAssetInfo](https://cloud.tencent.com/document/api/296/57282)
+* [DescribeAssetRecentMachineInfo](https://cloud.tencent.com/document/api/296/57281)
+* [SyncAssetScan](https://cloud.tencent.com/document/api/296/57280)
+
+新增数据结构：
+
+* [AssetKeyVal](https://cloud.tencent.com/document/api/296/19867#AssetKeyVal)
+
+
+
+## 弹性微服务(tem) 版本：2020-12-21
+
+### 第 6 次发布
+
+发布时间：2021-06-14 08:00:23
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改接口：
+
+* [DeployServiceV2](https://cloud.tencent.com/document/api/1371/55519)
+
+	* 新增入参：SettingConfs, EksService
+
+
+新增数据结构：
+
+* [EksService](https://cloud.tencent.com/document/api/1371/55421#EksService)
+* [MountedSettingConf](https://cloud.tencent.com/document/api/1371/55421#MountedSettingConf)
+
+
+
 # 历次版本更新如下：
 
 ## 活动防刷(aa) 版本：2020-02-24
@@ -15615,6 +15661,24 @@
 
 
 ## 主机安全(cwp) 版本：2018-02-28
+
+### 第 17 次发布
+
+发布时间：2021-06-14 08:10:35
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增接口：
+
+* [[DescribeAssetInfo](https://cloud.tencent.com/document/api/296/57282)](https://cloud.tencent.com/document/api/#/#)
+* [[DescribeAssetRecentMachineInfo](https://cloud.tencent.com/document/api/296/57281)](https://cloud.tencent.com/document/api/#/#)
+* [[SyncAssetScan](https://cloud.tencent.com/document/api/296/57280)](https://cloud.tencent.com/document/api/#/#)
+
+新增数据结构：
+
+* [[AssetKeyVal](https://cloud.tencent.com/document/api/296/19867#AssetKeyVal)](https://cloud.tencent.com/document/api/296/19867#[AssetKeyVal](https://cloud.tencent.com/document/api/296/19867#AssetKeyVal))
 
 ### 第 16 次发布
 
@@ -41492,6 +41556,26 @@
 
 
 ## 弹性微服务(tem) 版本：2020-12-21
+
+### 第 6 次发布
+
+发布时间：2021-06-14 08:00:23
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改接口：
+
+* [DeployServiceV2](https://cloud.tencent.com/document/api/1371/55519)
+
+	* 新增入参：SettingConfs, EksService
+
+
+新增数据结构：
+
+* [[EksService](https://cloud.tencent.com/document/api/1371/55421#EksService)](https://cloud.tencent.com/document/api/1371/55421#[EksService](https://cloud.tencent.com/document/api/1371/55421#EksService))
+* [[MountedSettingConf](https://cloud.tencent.com/document/api/1371/55421#MountedSettingConf)](https://cloud.tencent.com/document/api/1371/55421#[MountedSettingConf](https://cloud.tencent.com/document/api/1371/55421#MountedSettingConf))
 
 ### 第 5 次发布